Moved order creationg to the totals repo
